statystki

+++ src/signals1/signals/discrete/ImpulseInterface.java
@@ -23,7 +23,7 @@ public interface ImpulseInterface {
 
     SignalStats getStats();
 
-    int getDuration();
+    double getDuration();
     
     Histogram getHistogram(int numberOfIntervals);
 }
